add basic tests
